Use long instead of int for counter value
Change counter value type from int to long across domain model, events,
API response, and tests. Overflow-to-zero now occurs at Long.MAX_VALUE
instead of Integer.MAX_VALUE.

Resolves #3

+++ src/main/java/com/example/api/CounterEndpoint.java
@@ -12,7 +12,7 @@ import com.example.application.CounterEntity;
 @Acl(allow = @Acl.Matcher(principal = Acl.Principal.ALL))
 public class CounterEndpoint {
 
-  public record CounterResponse(String name, int value) {}
+  public record CounterResponse(String name, long value) {}
   public record ChangeNameRequest(String name) {}
 
   private final ComponentClient componentClient;
